+++ MyGUIEngine/src/MyGUI_List.cpp
@@ -193,7 +193,7 @@ namespace MyGUI
 
 	void List::notifyScrollChangePosition(MyGUI::WidgetPtr _sender, size_t _rel)
 	{
-		mOffsetTop = (_rel % mHeightLine);
+		mOffsetTop = ((int)_rel % mHeightLine);
 
 		// смещение с отрицательной стороны
 		int offset = 0 - mOffsetTop;
@@ -204,7 +204,7 @@ namespace MyGUI
 		}
 
 		// если индех изменился, то перерисовываем линии
-		int top = (_rel / mHeightLine);
+		int top = ((int)_rel / mHeightLine);
 		if (top != mTopIndex) {
 			mTopIndex = top;
 			_redrawItemRange();
